Fix experiment_utils.py: handling methods with the same name and signature
The original implementation used "class file + method name + method signature" as a method identifier so that it was possible that different methods in a class file have the same identifiers if they are defined in different classes. In those special cases, the depth of the commit can be turned to a negative value when adjusting the commit depth after discarding the style-change commits. So I added the line numbers to the method identifiers to make sure that different methods have different identifiers.